+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 * **terra-checkbox-group** new component which wraps `<terra-multi-check-box>` to separate available checkbox definition from ngModel
 * **terra-form** added support for `<terra-checkbox-group>`
 
+### Bug Fixes
+* **terra-form-entry** fixed initialization with defaultValue
+
 <a name="3.2.0"></a>
 # 3.2.0 (23.01.2019)